refactor panel->edit
