+++ components/form-builder/edit/ElementPanel.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useTranslation } from "next-i18next";
 
 import { FormElementWithIndex } from "../types";
-import { useTemplateStore, useModalStore } from "../store/";
+import { useTemplateStore, useModalStore } from "../store";
 import { PanelActions, ModalButton, ModalForm, PanelBody } from "./index";
 import { Button } from "../shared";
