+++ components/form-builder/edit/PanelActionsLocked.tsx
