+++ components/form-builder/edit/tests/ElementPanel.test.js
